# docs/openapi/compute-service.yaml
...
components:
  schemas:
    Corev1ConfigMapEnvSource:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      type: object
    Corev1ConfigMapKeySelector:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        key:
          type: string
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      - key
      type: object
    Corev1EnvFromSource:
      properties:
        configMapRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ConfigMapEnvSource'
          nullable: true
        prefix:
          type: string
        secretRef:
          allOf:
          - $ref: '#/components/schemas/Corev1SecretEnvSource'
          nullable: true
      type: object
    Corev1EnvVar:
      properties:
        name:
          type: string
        value:
          type: string
        valueFrom:
          allOf:
          - $ref: '#/components/schemas/Corev1EnvVarSource'
          nullable: true
      required:
      - name
      type: object
    Corev1EnvVarSource:
      properties:
        configMapKeyRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ConfigMapKeySelector'
          nullable: true
        fieldRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ObjectFieldSelector'
          nullable: true
        fileKeyRef:
          allOf:
          - $ref: '#/components/schemas/Corev1FileKeySelector'
          nullable: true
        resourceFieldRef:
          allOf:
          - $ref: '#/components/schemas/Corev1ResourceFieldSelector'
          nullable: true
        secretKeyRef:
          allOf:
          - $ref: '#/components/schemas/Corev1SecretKeySelector'
          nullable: true
      type: object
    Corev1FileKeySelector:
      properties:
        key:
          type: string
        optional:
          nullable: true
          type: boolean
        path:
          type: string
        volumeName:
          type: string
      required:
      - key
      - path
      - volumeName
      type: object
    Corev1LocalObjectReference:
      properties:
        name:
          type: string
      type: object
    Corev1ObjectFieldSelector:
      properties:
        apiVersion:
          type: string
        fieldPath:
          type: string
      required:
      - fieldPath
      type: object
    Corev1ResourceClaim:
      properties:
        name:
          type: string
        request:
          type: string
      required:
      - name
      type: object
    Corev1ResourceFieldSelector:
      properties:
        containerName:
          type: string
        divisor:
          description: Kubernetes resource.Quantity (e.g. "500m", "2Gi", "4").
          type: string
        resource:
          type: string
      required:
      - resource
      type: object
    Corev1ResourceRequirements:
      properties:
        claims:
          items:
            $ref: '#/components/schemas/Corev1ResourceClaim'
          type: array
        limits:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
        requests:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
      type: object
    Corev1SecretEnvSource:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      type: object
    Corev1SecretKeySelector:
      properties:
        LocalObjectReference:
          $ref: '#/components/schemas/Corev1LocalObjectReference'
        key:
          type: string
        optional:
          nullable: true
          type: boolean
      required:
      - LocalObjectReference
      - key
      type: object
    Corev1Toleration:
      properties:
        effect:
          type: string
        key:
          type: string
        operator:
          type: string
        tolerationSeconds:
          format: int64
          nullable: true
          type: integer
        value:
          type: string
      type: object
    JobCreateInput:
      properties:
        backend:
          allOf:
          - $ref: '#/components/schemas/MLJobBackendSpec'
          nullable: true
        description:
          type: string
        displayName:
          type: string
        name:
          maxLength: 40
          minLength: 3
          pattern: ^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])?$
          type: string
        quota:
          type: string
        resourceUnitId:
          format: uuid
          type: string
        roles:
          items:
            $ref: '#/components/schemas/MLJobRoleSpec'
          minItems: 1
          type: array
        runPolicy:
          allOf:
          - $ref: '#/components/schemas/MLJobRunPolicySpec'
          nullable: true
      required:
      - name
      - quota
      - resourceUnitId
      - roles
      type: object
    JobList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/JobView'
          type: array
        total:
          format: int64
          type: integer
      required:
      - items
      - total
      type: object
    JobView:
      properties:
        createdAt:
          format: date-time
          type: string
        description:
          type: string
        displayName:
          type: string
        finishedAt:
          format: date-time
          nullable: true
          type: string
        id:
          format: uuid
          type: string
        message:
          type: string
        name:
          type: string
        namespace:
          type: string
        ownerUser:
          type: string
        spec:
          $ref: '#/components/schemas/MLJobSpec'
        startedAt:
          format: date-time
          nullable: true
          type: string
        status:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - name
      - namespace
      - spec
      - status
      - updatedAt
      type: object
    MLJobBackendSpec:
      properties:
        config:
          description: Embedded Kubernetes resource (free-form JSON).
          nullable: true
          type: object
        engine:
          type: string
        name:
          type: string
      required:
      - engine
      - name
      type: object
    MLJobPodTemplateSubset:
      properties:
        args:
          items:
            type: string
          type: array
        command:
          items:
            type: string
          type: array
        env:
          items:
            $ref: '#/components/schemas/Corev1EnvVar'
          type: array
        envFrom:
          items:
            $ref: '#/components/schemas/Corev1EnvFromSource'
          type: array
        image:
          type: string
        imagePullPolicy:
          type: string
        resources:
          $ref: '#/components/schemas/Corev1ResourceRequirements'
        workingDir:
          type: string
      required:
      - image
      type: object
    MLJobRoleSpec:
      properties:
        name:
          type: string
        replicas:
          format: int32
          type: integer
        restartPolicy:
          type: string
        template:
          $ref: '#/components/schemas/MLJobPodTemplateSubset'
      required:
      - name
      - replicas
      - template
      type: object
    MLJobRunPolicySpec:
      properties:
        activeDeadlineSeconds:
          format: int64
          nullable: true
          type: integer
        backoffLimit:
          format: int32
          nullable: true
          type: integer
        suspend:
          type: boolean
        ttlSecondsAfterFinished:
          format: int32
          nullable: true
          type: integer
      type: object
    MLJobSchedulingSpec:
      properties:
        nodeSelector:
          additionalProperties:
            type: string
          type: object
        priorityClass:
          type: string
        quota:
          type: string
        tolerations:
          items:
            $ref: '#/components/schemas/Corev1Toleration'
          type: array
      required:
      - quota
      type: object
    MLJobSpec:
      properties:
        backend:
          $ref: '#/components/schemas/MLJobBackendSpec'
        roles:
          items:
            $ref: '#/components/schemas/MLJobRoleSpec'
          type: array
        runPolicy:
          $ref: '#/components/schemas/MLJobRunPolicySpec'
        scheduling:
          $ref: '#/components/schemas/MLJobSchedulingSpec'
      required:
      - backend
      - roles
      - scheduling
      type: object
...
    Problem:
      properties:
        code:
          description: Discrete business error class.
          enum:
          - validation_failed
          - not_found
          - conflict
          - precondition_failed
          - unauthorized
          - forbidden
          - service_unavailable
          - internal_error
          - quota_exceeded
          type: string
        detail:
          type: string
        details:
          additionalProperties: {}
          type: object
        instance:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
      required:
      - code
      - status
      - title
      - type
      type: object
...
info:
  description: REST API for resource pools, resource units, jobs, and services. Partitioned
    by Kubernetes namespace. RFC7807 Problem responses on errors.
  title: AxisML Compute Service API
  version: 0.1.0
openapi: 3.0.3
paths:
  /api/v1/namespaces/{namespace}/jobs:
    get:
      operationId: listJobs
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Page size (1–200, default 50).
        in: query
        name: limit
        schema:
          format: int32
          type: integer
      - description: Number of items to skip.
        in: query
        name: offset
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobList'
          description: Page.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: List jobs in a namespace
      tags:
      - jobs
    post:
      operationId: createJob
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobView'
          description: Created.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Submit an MLJob
      tags:
      - jobs
  /api/v1/namespaces/{namespace}/jobs/{job}:
    delete:
      operationId: deleteJob
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Job name.
        in: path
        name: job
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted (idempotent).
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Delete job
      tags:
      - jobs
    get:
      operationId: getJob
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Job name.
        in: path
        name: job
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobView'
          description: Job.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Get job
      tags:
      - jobs
  /api/v1/namespaces/{namespace}/jobs/{job}/cancel:
    post:
      operationId: cancelJob
      parameters:
      - description: Kubernetes namespace.
        in: path
        name: namespace
        required: true
        schema:
          type: string
      - description: Job name.
        in: path
        name: job
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobView'
          description: Cancelled job.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Cancel a running job
      tags:
      - jobs
...
servers:
- description: Same-origin
  url: /
...